requirements.txt
anthropic>=0.40.0
pillow>=10.0.0
pillow-heif>=0.18.0
tenacity>=8.2.0
tqdm>=4.66.0
